requirements.txt
playwright
pandas
psycopg2
python-dotenv
sqlalchemy
jupyterlab
lightgbm
scikit-learn
shap
matplotlib
seaborn
pydantic
fastapi
uvicorn[standard]
langgraph
pytest
pytest-asyncio
